tabs, old libgdx version, default constructor
