+++ android/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
-        package="net.spookygames.gdx.notifications">
+        package="net.spookygames.gdx.notifications.android">
 </manifest>
