+++ build.gradle
@@ -18,7 +18,7 @@ allprojects {
     version = '0.0.2'
 
     ext {
-        gdxVersion = '1.9.9'
+        gdxVersion = '1.5.5'
         libName = 'gdx-notifications'
     }
