+++ html-browser/src/main/java/net/spookygames/gdx/notifications/HtmlNotificationHandler.java
@@ -29,126 +29,126 @@ import com.google.gwt.core.client.JavaScriptObject;
 
 public class HtmlNotificationHandler implements NotificationHandler {
 
-    private boolean supported;
-    private boolean hasPermission = false;
-    private boolean permissionRequested = false;
+	private boolean supported;
+	private boolean hasPermission = false;
+	private boolean permissionRequested = false;
 
-    private Array<NotificationJS> notifications = new Array<NotificationJS>();
+	private Array<NotificationJS> notifications = new Array<NotificationJS>();
 
-    private String icon;
+	private String icon;
 
-    public HtmlNotificationHandler() {
-        this.supported = isSupported();
-        if (this.supported) {
-            this.hasPermission = hasPermission();
-        }
-    }
+	public HtmlNotificationHandler() {
+		this.supported = isSupported();
+		if (this.supported) {
+			this.hasPermission = hasPermission();
+		}
+	}
 
-    @Override
-    public void showNotification(NotificationParameters parameters) {
-        if (this.supported) {
-            if (this.hasPermission) {
-                this.notifications.add(showNotificationJSNI(parameters, this.icon));
-            } else {
-                if (!this.permissionRequested) {
-                    this.permissionRequested = true;
-                    this.requestPermission(new NotificationPermissionResult(parameters));
-                }
-            }
-        }
-    }
+	@Override
+	public void showNotification(NotificationParameters parameters) {
+		if (this.supported) {
+			if (this.hasPermission) {
+				this.notifications.add(showNotificationJSNI(parameters, this.icon));
+			} else {
+				if (!this.permissionRequested) {
+					this.permissionRequested = true;
+					this.requestPermission(new NotificationPermissionResult(parameters));
+				}
+			}
+		}
+	}
 
-    @Override
-    public void hideNotification(NotificationParameters parameters) {
-        if (this.supported) {
-            for (NotificationJS notification : this.notifications) {
-                if (Integer.valueOf(notification.getTag()) == parameters.getId()) {
-                    notification.close();
-                    this.notifications.removeValue(notification, true);
-                    break;
-                }
-            }
-        }
-    }
+	@Override
+	public void hideNotification(NotificationParameters parameters) {
+		if (this.supported) {
+			for (NotificationJS notification : this.notifications) {
+				if (Integer.valueOf(notification.getTag()) == parameters.getId()) {
+					notification.close();
+					this.notifications.removeValue(notification, true);
+					break;
+				}
+			}
+		}
+	}
 
-    /**
-     * Sets the icon for all notifications
-     * @param icon path to the icon
-     * @param type internal: icon from assets, external/absolute: icon from external domain yourdomain.com/picture.png
-     */
-    public void setIcon(String icon, Files.FileType type) {
-        switch (type) {
-            case Internal:
-                this.icon = "assets/" + icon;
-                break;
-            case External:
-            case Absolute:
-                this.icon = icon;
-                break;
-        }
-    }
+	/**
+	 * Sets the icon for all notifications
+	 * @param icon path to the icon
+	 * @param type internal: icon from assets, external/absolute: icon from external domain yourdomain.com/picture.png
+	 */
+	public void setIcon(String icon, Files.FileType type) {
+		switch (type) {
+			case Internal:
+				this.icon = "assets/" + icon;
+				break;
+			case External:
+			case Absolute:
+				this.icon = icon;
+				break;
+		}
+	}
 
-    private native NotificationJS showNotificationJSNI(NotificationParameters parameters, String icon) /*-{
-        return new $wnd.Notification(parameters.@net.spookygames.gdx.notifications.NotificationParameters::getTitle()(), {
-            tag: parameters.@net.spookygames.gdx.notifications.NotificationParameters::getId()(),
-            body: parameters.@net.spookygames.gdx.notifications.NotificationParameters::getText()(),
-            icon: icon
-        });
-    }-*/;
+	private native NotificationJS showNotificationJSNI(NotificationParameters parameters, String icon) /*-{
+		return new $wnd.Notification(parameters.@net.spookygames.gdx.notifications.NotificationParameters::getTitle()(), {
+			tag: parameters.@net.spookygames.gdx.notifications.NotificationParameters::getId()(),
+			body: parameters.@net.spookygames.gdx.notifications.NotificationParameters::getText()(),
+			icon: icon
+		});
+	}-*/;
 
-    private native boolean isSupported() /*-{
-        return "Notification" in $wnd;
-    }-*/;
+	private native boolean isSupported() /*-{
+		return "Notification" in $wnd;
+	}-*/;
 
-    private native boolean hasPermission() /*-{
-        return $wnd.Notification.permission === "granted";
-    }-*/;
+	private native boolean hasPermission() /*-{
+		return $wnd.Notification.permission === "granted";
+	}-*/;
 
-    private native void requestPermission(NotificationPermissionResult result) /*-{
-        $wnd.Notification.requestPermission(function (permissionStatus) {
-            if (permissionStatus === 'granted') {
-                result.@net.spookygames.gdx.notifications.HtmlNotificationHandler.NotificationPermissionResult::granted()();
-            } else if (permissionStatus === 'denied') {
-                result.@net.spookygames.gdx.notifications.HtmlNotificationHandler.NotificationPermissionResult::denied()();
-            } else if (permissionStatus === 'default') {
-                result.@net.spookygames.gdx.notifications.HtmlNotificationHandler.NotificationPermissionResult::unknown()();
-            }
-        });
-    }-*/;
+	private native void requestPermission(NotificationPermissionResult result) /*-{
+		$wnd.Notification.requestPermission(function (permissionStatus) {
+			if (permissionStatus === 'granted') {
+				result.@net.spookygames.gdx.notifications.HtmlNotificationHandler.NotificationPermissionResult::granted()();
+			} else if (permissionStatus === 'denied') {
+				result.@net.spookygames.gdx.notifications.HtmlNotificationHandler.NotificationPermissionResult::denied()();
+			} else if (permissionStatus === 'default') {
+				result.@net.spookygames.gdx.notifications.HtmlNotificationHandler.NotificationPermissionResult::unknown()();
+			}
+		});
+	}-*/;
 
-    private static class NotificationJS extends JavaScriptObject {
-        protected NotificationJS() {
-        }
+	private static class NotificationJS extends JavaScriptObject {
+		protected NotificationJS() {
+		}
 
-        native final String getTag() /*-{
-            return this.tag;
-        }-*/;
+		native final String getTag() /*-{
+			return this.tag;
+		}-*/;
 
-        native final void close() /*-{
-            this.close();
-        }-*/;
-    }
+		native final void close() /*-{
+			this.close();
+		}-*/;
+	}
 
-    private class NotificationPermissionResult {
+	private class NotificationPermissionResult {
 
-        NotificationParameters parameters;
+		NotificationParameters parameters;
 
-        NotificationPermissionResult(NotificationParameters parameters) {
-            this.parameters = parameters;
-        }
+		NotificationPermissionResult(NotificationParameters parameters) {
+			this.parameters = parameters;
+		}
 
-        public void granted() {
-            hasPermission = true;
-            showNotification(this.parameters);
-        }
+		public void granted() {
+			hasPermission = true;
+			showNotification(this.parameters);
+		}
 
-        public void denied() {
-            hasPermission = false;
-        }
+		public void denied() {
+			hasPermission = false;
+		}
 
-        public void unknown() {
-            permissionRequested = false;
-            showNotification(this.parameters);
-        }
-    }
+		public void unknown() {
+			permissionRequested = false;
+			showNotification(this.parameters);
+		}
+	}
 }
